feat: support import of external CSS
Base on our current version of TailwindCSS and NextJS we are still on an
older version of PostCSS. The new postcss-import plugin relies on
PostCSS v8 but we need to stay compatible with v7. Therefore, we are
installing postcss-import v12 and not the newer v14.

Both packages should be updated to later version and the postcss-import
plugin shold be updated as well.

https://tailwindcss.com/docs/installation#post-css-7-compatibility-build
https://github.com/vercel/next.js/issues/17236