fix: avoid render various amount of element descriptions
The previous implementation hydrated the element descriptions once for
each indicator page and supplied the hydrated MDX sources to child
components as props. The amount of element descriptions varies across
indicators which led to the situation that on each render a different
amount of indicators were rendered. This is a violation of the rules of
hooks and caused an error.

Element descriptions are hydrated once for each place where they appear.
This certainly isn't nearly as performant as the previous version but
doesn't throw an error. Let's see if it's good enough.